Struct wasmer_registry::config::MultiRegistry

pub struct MultiRegistry {
    pub active_registry: String,
    pub tokens: Vec<RegistryLogin>,
}

Struct to store login tokens for multiple registry URLs inside of the wasmer.toml configuration file

Fields
active_registry: String

Currently active registry

tokens: Vec<RegistryLogin>

Map from “RegistryUrl” to “LoginToken”, in order to be able to be able to easily switch between registries

Implementations

impl MultiRegistry

pub fn clear_current_registry_token(&mut self)

Gets the current (active) registry URL

pub fn get_graphql_url(&self) -> String

pub fn get_current_registry(&self) -> String

Gets the current (active) registry URL

pub fn set_current_registry(&mut self, registry: &str)

Sets the current (active) registry URL

pub fn get_login_token_for_registry(&self, registry: &str) -> Option<String>

Returns the login token for the registry

pub fn set_login_token_for_registry(
&mut self,
registry: &str,
token: &str,
update_current_registry: UpdateRegistry
)

Sets the login token for the registry URL