Function wasmer_registry::graphql::execute_query_modifier_inner_check_json

pub fn execute_query_modifier_inner_check_json<V, F>(
    registry_url: &str,
    login_token: &str,
    query: &QueryBody<V>,
    timeout: Option<Duration>,
    form_modifier: F
) -> Result<()> where
    V: Serialize,
    F: FnOnce(Form) -> Form, 

This function is being used to “ping” the registry (see test_if_registry_present and see whether the response is valid JSON, it doesn’t check the response itself, since the response format might change