Function wasmer_registry::graphql::proxy::maybe_set_up_proxy_inner

fn maybe_set_up_proxy_inner() -> Result<Option<Proxy>>

Tries to set up a proxy

This function reads from wasmer config’s proxy.url first, then checks ALL_PROXY, HTTPS_PROXY, and HTTP_PROXY environment variables, in both upper case and lower case, in that order.

If a proxy is specified in wasmer config’s proxy.url, it is assumed to be a general proxy

A return value of Ok(None) means that there was no attempt to set up a proxy, Ok(Some(proxy)) means that the proxy was set up successfully, and Err(e) that there was a failure while attempting to set up the proxy.