Struct wasmer_registry::package::builder::Publish

pub struct Publish {
    pub registry: Option<String>,
    pub dry_run: bool,
    pub quiet: bool,
    pub package_name: Option<String>,
    pub version: Option<Version>,
    pub token: Option<String>,
    pub no_validate: bool,
    pub package_path: Option<String>,
}

CLI options for the wasmer publish command

Fields
registry: Option<String>

Registry to publish to

dry_run: bool

Run the publish logic without sending anything to the registry server

quiet: bool

Run the publish command without any output

package_name: Option<String>

Override the package of the uploaded package in the wasmer.toml

version: Option<Version>

Override the package version of the uploaded package in the wasmer.toml

token: Option<String>

Override the token (by default, it will use the current logged in user)

no_validate: bool

Skip validation of the uploaded package

package_path: Option<String>

Directory containing the wasmer.toml (defaults to current root dir)

Implementations

impl Publish

pub fn execute(&self) -> Result<(), Error>

Executes wasmer publish