Struct wasmer_registry::package::Package

pub struct Package {
    pub namespace: String,
    pub name: String,
    pub version: Option<String>,
}

Fields
namespace: Stringname: Stringversion: Option<String>
Implementations

impl Package

pub fn already_installed(&self, wasmer_dir: &Path) -> Option<PathBuf>

Checks whether the package is already installed, if yes, returns the path to the root dir

pub fn is_url_already_installed(url: &Url, wasmer_dir: &Path) -> Option<PathBuf>

Checks if the URL is already installed, note that {url}@{version} and {url} are treated the same

pub fn hash_url(url: &str) -> String

Returns the hash of the URL with a maximum of 128 bytes length (necessary for not erroring on filesystem limitations)

pub fn unhash_url(hashed: &str) -> String

Returns the hash of the URL with a maximum of 64 bytes length

pub fn get_hash(&self, registry: &str) -> String

Returns the hash of the package URL without the version (because the version is encoded as @version and isn’t part of the hash itself)

fn get_url_without_version(&self, registry: &str) -> Result<String, Error>

pub fn file(&self) -> String

Returns the filename for this package

pub fn package(&self) -> String

Returns the {namespace}/{name} package name

pub fn url(&self, registry: &str) -> Result<Url, Error>

Returns the full URL including the version for this package

pub fn get_path(&self, wasmer_dir: &Path) -> Result<PathBuf, Error>

Returns the path to the installation directory. Does not check whether the installation directory already exists.