Module wasmer_registry::queries

Low-level GraphQL queries used by this crate.

If possible, users should prefer the high-level functions exposed under crate.

This module is primarily used in combination with crate::graphql::execute_query() as an “escape hatch” for accessing information that may not be exposed via the high-level functions.

Backwards Compatibility

Queries won’t be deleted or have breaking changes to their inputs during patch releases, however new fields may be added to the response types generated by graphql_client at any time.

Users should treat all response types as if they had the #[non_exhaustive] attribute.

Modules

get_bindings_query

get_package_by_command_query

get_package_version_query

test_if_registry_present

who_am_i_query

Structs

GetBindingsQuery

GetPackageByCommandQuery

GetPackageVersionQuery

TestIfRegistryPresent

WhoAmIQuery

Constants

SCHEMA

The GraphQL schema exposed by the WAPM backend.