Struct wasmer_registry::wasmer_env::WasmerEnv

pub struct WasmerEnv {
    wasmer_dir: PathBuf,
    registry: Option<Registry>,
    cache_dir: Option<PathBuf>,
    token: Option<String>,
}

Command-line flags for determining the local “Wasmer Environment”.

This is where you access $WASMER_DIR, the $WASMER_DIR/wasmer.toml config file, and specify the current registry.

Fields
wasmer_dir: PathBuf

Set Wasmer’s home directory

registry: Option<Registry>

The registry to fetch packages from (inferred from the environment by default)

cache_dir: Option<PathBuf>

The directory cached artefacts are saved to.

token: Option<String>

The API token to use when communicating with the registry (inferred from the environment by default)

Implementations

impl WasmerEnv

pub fn new( wasmer_dir: PathBuf, registry: Option<Registry>, token: Option<String>, cache_dir: Option<PathBuf> ) -> Self

pub fn registry_endpoint(&self) -> Result<Url, Error>

Get the GraphQL endpoint used to query the registry.

pub fn config(&self) -> Result<WasmerConfig, Error>

Load the current Wasmer config.

pub fn dir(&self) -> &Path

The directory all Wasmer artifacts are stored in.

pub fn cache_dir(&self) -> PathBuf

The directory all cached artifacts should be saved to.

pub fn token(&self) -> Option<String>

The API token for the active registry.