Struct wasmer_registry::config::WasmerConfig

pub struct WasmerConfig {
    pub telemetry_enabled: bool,
    pub update_notifications_enabled: bool,
    pub registry: MultiRegistry,
    pub proxy: Proxy,
}

Fields
telemetry_enabled: bool

Whether or not telemetry is enabled.

update_notifications_enabled: bool

Whether or not updated notifications are enabled.

registry: MultiRegistry

The registry that wasmer will connect to.

proxy: Proxy

The proxy to use when connecting to the Internet.

Implementations

impl WasmerConfig

pub(crate) const ENV_VAR_WASMER_REGISTRY_TOKEN: &str = "WASMER_TOKEN"

pub(crate) const ENV_VAR_WASMER_REGISTRY_TOKEN_LEGACY: &str = "WAPM_REGISTRY_TOKEN"

pub fn save<P: AsRef<Path>>(&self, to: P) -> Result<()>

Save the config to a file

pub fn from_file(wasmer_dir: &Path) -> Result<Self, String>

pub fn get_wasmer_dir() -> Result<PathBuf, String>

Creates and returns the WASMER_DIR directory (or $HOME/.wasmer as a fallback)

pub fn from_env() -> Result<Self, Error>

Load the config based on environment variables and default config file locations.

pub fn get_current_dir() -> Result<PathBuf>

pub fn get_file_location(wasmer_dir: &Path) -> PathBuf

pub fn get_database_file_path(wasmer_dir: &Path) -> PathBuf